Testing D(k) with k in the full BZ
Former-commit-id: 6a44596385c2a5834cec47ffa07648daa83d0438
Former-commit-id: 35a8e748940fa242c2153305588303416578d72e

+++ shared/common/src/32_util/m_symtk.F90
@@ -3448,7 +3448,7 @@ subroutine sym_order(rot, tnons, n, isproper, trans, msg, ierr)
  end do
 
  ! Make sure trans_dp is integer.
- trans = int(trans_dp)
+ trans = nint(trans_dp)
  if (.not. isinteger(trans_dp, tol=tol5)) then
    ierr = 2
    msg = sjoin("T = [I + rot + rot^2 + ... + rot^(n-1)] . tnons is not integer:", ltoa(trans_dp))
